Create new finger joint construct
Use functional concepts to create finger joint

Eventually add:
- A and B fingers, with rounded tips
- A and B slots, to be subtracted from other parts

In the case of the fingers, the entire area of the joint should be
removed. This may not always be needed but with help reduce the
mental overhead of needing to explicitly add/remove material from
a joint so the focus can remain on the interfacing of parts.