Add a rough shape of the receiving plate
In testing the magsafe charging adapter, if the stand has the
device slot straight in and out, the adapter will be pulled from
the device. The receiving plate should be more of a backing.

An alternative idea is that through a rotating joint or other
mechanism the device and its cradle "lock in" to a charging
position, which could also be compatible with the microswitch
trigger idea.